fix(cadastro): improve the responsiveness of the registration activation screen and avoid overflow
To improve the responsiveness of the registration activation screen and avoid overflow. I changed
the method of getting the current width/height of the screen since window.screen.width/height
doesn't represent the width or height of the browser. With innerWidth and innerHeight the confetti
animation will be responsive as expected.

fix #898