refactor(request builder): create a request builder and use it in `contents` integration tests
The `RequestBuilder` is a way to facilitate the writing of tests, reducing the repetition of user
creation, activation, obtaining the token, insertion or removal of `feature`, and finally `fetch`.
The goal is to make tests easier to write, while still being clear about what is being done, so it
remains clear how the test works when you read it.