refactor(content): in `creditOrDebitTabCoins`, get only the required fields from parent
In `creditOrDebitTabCoins`, changing the order of `if`s prevents a database query when the content
is not relevant. Furthermore, searching the parent content directly with SQL allows to obtain only
the desired field (`owner_id`), instead of bringing up all the fields from the table and also
calculating the content's TabCoins.